feat: Add product data and components for cart functionality in example page
This commit creats cart example page and adds the necessary code changes to support the cart functionality in the web store. It includes the addition of the `products.server.ts` file, which contains the product data for the cart. It also introduces two new components: `ProductItemCard` and `CartItemCard`, which are responsible for rendering the product items and cart items respectively. Additionally, a new `NumberInput` component is added to handle quantity input.